+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to Rust's notion of
 
 ## [Unreleased]
 
+### Added
+- `impl subtle::ConstantTimeEq for note::Nullifier`
+- `impl Clone for note_encryption::CompactAction`
+- `note_encryption::CompactAction::cmx`
+
 ## [0.7.0] - 2024-01-26
 ### Licensing
 - The license for this crate is now "MIT OR Apache-2.0". The license